signed -> unsigned

+++ Marlin/src/lcd/e_paper/epd/epaper/GxEPD2_371.cpp
@@ -26,32 +26,29 @@ void GxEPD2_371::writeScreenBuffer(uint8_t value) {
   _initial_write = false;  // initial full screen buffer clean done
   if (!_using_partial_mode) _Init_Part();
   _writeCommand(0x13);     // set current
-  for (uint32_t i = 0; i < uint32_t(WIDTH) * uint32_t(HEIGHT) / 8; i++)
-  {
+  for (uint16_t i = 0; i < uint16_t(WIDTH) * uint16_t(HEIGHT) / 8; i++) {
     _writeData(value);
   }
-  if (_initial_refresh)
-  {
+  if (_initial_refresh) {
     _writeCommand(0x10);   // preset previous
-    for (uint32_t i = 0; i < uint32_t(WIDTH) * uint32_t(HEIGHT) / 8; i++)
-    {
+    for (uint16_t i = 0; i < uint16_t(WIDTH) * uint16_t(HEIGHT) / 8; i++) {
       _writeData(0xFF);    // 0xFF is white
     }
   }
 }
 
-void GxEPD2_371::writeImage(const uint8_t bitmap[], int16_t x, int16_t y, int16_t w, int16_t h, bool invert, bool mirror_y, bool pgm) {
+void GxEPD2_371::writeImage(const uint8_t bitmap[], uint16_t x, uint16_t y, uint16_t w, uint16_t h, bool invert, bool mirror_y, bool pgm) {
   if (_initial_write) writeScreenBuffer(); // initial full screen buffer clean
   delay(1);                // yield() to avoid WDT on ESP8266 and ESP32
-  int16_t wb = (w + 7) / 8; // width bytes, bitmaps are padded
+  uint16_t wb = (w + 7) / 8; // width bytes, bitmaps are padded
   x -= x % 8;              // byte boundary
   w = wb * 8;              // byte boundary
-  int16_t x1 = x < 0 ? 0 : x; // limit
-  int16_t y1 = y < 0 ? 0 : y; // limit
-  int16_t w1 = x + w < int16_t(WIDTH) ? w : int16_t(WIDTH) - x; // limit
-  int16_t h1 = y + h < int16_t(HEIGHT) ? h : int16_t(HEIGHT) - y; // limit
-  int16_t dx = x1 - x;
-  int16_t dy = y1 - y;
+  uint16_t x1 = x < 0 ? 0 : x; // limit
+  uint16_t y1 = y < 0 ? 0 : y; // limit
+  uint16_t w1 = x + w < uint16_t(WIDTH) ? w : uint16_t(WIDTH) - x; // limit
+  uint16_t h1 = y + h < uint16_t(HEIGHT) ? h : uint16_t(HEIGHT) - y; // limit
+  uint16_t dx = x1 - x;
+  uint16_t dy = y1 - y;
   w1 -= dx;
   h1 -= dy;
   if ((w1 <= 0) || (h1 <= 0)) return;
@@ -59,11 +56,11 @@ void GxEPD2_371::writeImage(const uint8_t bitmap[], int16_t x, int16_t y, int16_
   _writeCommand(0x91); // partial in
   _setPartialRamArea(x1, y1, w1, h1);
   _writeCommand(0x13);
-  for (int16_t i = 0; i < h1; i++) {
-    for (int16_t j = 0; j < w1 / 8; j++) {
+  for (uint16_t i = 0; i < h1; i++) {
+    for (uint16_t j = 0; j < w1 / 8; j++) {
       uint8_t data;
       // use wb, h of bitmap for index!
-      int16_t idx = mirror_y ? j + dx / 8 + ((h - 1 - (i + dy))) * wb : j + dx / 8 + (i + dy) * wb;
+      uint16_t idx = mirror_y ? j + dx / 8 + ((h - 1 - (i + dy))) * wb : j + dx / 8 + (i + dy) * wb;
       if (pgm) {
         #if defined(__AVR) || defined(ESP8266) || defined(ESP32)
           data = pgm_read_byte(&bitmap[idx]);
@@ -82,25 +79,25 @@ void GxEPD2_371::writeImage(const uint8_t bitmap[], int16_t x, int16_t y, int16_
   delay(1);                // yield() to avoid WDT on ESP8266 and ESP32
 }
 
-void GxEPD2_371::writeImagePart(const uint8_t bitmap[], int16_t x_part, int16_t y_part, int16_t w_bitmap, int16_t h_bitmap,
-                                int16_t x, int16_t y, int16_t w, int16_t h, bool invert, bool mirror_y, bool pgm) {
+void GxEPD2_371::writeImagePart(const uint8_t bitmap[], uint16_t x_part, uint16_t y_part, uint16_t w_bitmap, uint16_t h_bitmap,
+                                uint16_t x, uint16_t y, uint16_t w, uint16_t h, bool invert, bool mirror_y, bool pgm) {
   if (_initial_write) writeScreenBuffer(); // initial full screen buffer clean
   delay(1);                // yield() to avoid WDT on ESP8266 and ESP32
   if ((w_bitmap < 0) || (h_bitmap < 0) || (w < 0) || (h < 0)) return;
   if ((x_part < 0) || (x_part >= w_bitmap)) return;
   if ((y_part < 0) || (y_part >= h_bitmap)) return;
-  int16_t wb_bitmap = (w_bitmap + 7) / 8; // width bytes, bitmaps are padded
+  uint16_t wb_bitmap = (w_bitmap + 7) / 8; // width bytes, bitmaps are padded
   x_part -= x_part % 8;    // byte boundary
   w = w_bitmap - x_part < w ? w_bitmap - x_part : w; // limit
   h = h_bitmap - y_part < h ? h_bitmap - y_part : h; // limit
   x -= x % 8;              // byte boundary
   w = 8 * ((w + 7) / 8);   // byte boundary, bitmaps are padded
-  int16_t x1 = x < 0 ? 0 : x; // limit
-  int16_t y1 = y < 0 ? 0 : y; // limit
-  int16_t w1 = x + w < int16_t(WIDTH) ? w : int16_t(WIDTH) - x; // limit
-  int16_t h1 = y + h < int16_t(HEIGHT) ? h : int16_t(HEIGHT) - y; // limit
-  int16_t dx = x1 - x;
-  int16_t dy = y1 - y;
+  uint16_t x1 = x < 0 ? 0 : x; // limit
+  uint16_t y1 = y < 0 ? 0 : y; // limit
+  uint16_t w1 = x + w < uint16_t(WIDTH) ? w : uint16_t(WIDTH) - x; // limit
+  uint16_t h1 = y + h < uint16_t(HEIGHT) ? h : uint16_t(HEIGHT) - y; // limit
+  uint16_t dx = x1 - x;
+  uint16_t dy = y1 - y;
   w1 -= dx;
   h1 -= dy;
   if ((w1 <= 0) || (h1 <= 0)) return;
@@ -108,11 +105,11 @@ void GxEPD2_371::writeImagePart(const uint8_t bitmap[], int16_t x_part, int16_t 
   _writeCommand(0x91);     // partial in
   _setPartialRamArea(x1, y1, w1, h1);
   _writeCommand(0x13);
-  for (int16_t i = 0; i < h1; i++) {
-    for (int16_t j = 0; j < w1 / 8; j++) {
+  for (uint16_t i = 0; i < h1; i++) {
+    for (uint16_t j = 0; j < w1 / 8; j++) {
       uint8_t data;
       // use wb_bitmap, h_bitmap of bitmap for index!
-      int16_t idx = mirror_y ? x_part / 8 + j + dx / 8 + ((h_bitmap - 1 - (y_part + i + dy))) * wb_bitmap : x_part / 8 + j + dx / 8 + (y_part + i + dy) * wb_bitmap;
+      uint16_t idx = mirror_y ? x_part / 8 + j + dx / 8 + ((h_bitmap - 1 - (y_part + i + dy))) * wb_bitmap : x_part / 8 + j + dx / 8 + (y_part + i + dy) * wb_bitmap;
       if (pgm) {
         #if defined(__AVR) || defined(ESP8266) || defined(ESP32)
           data = pgm_read_byte(&bitmap[idx]);
@@ -131,49 +128,49 @@ void GxEPD2_371::writeImagePart(const uint8_t bitmap[], int16_t x_part, int16_t 
   delay(1);                // yield() to avoid WDT on ESP8266 and ESP32
 }
 
-void GxEPD2_371::writeImage(const uint8_t* black, const uint8_t* color, int16_t x, int16_t y, int16_t w, int16_t h, bool invert, bool mirror_y, bool pgm) {
+void GxEPD2_371::writeImage(const uint8_t* black, const uint8_t* color, uint16_t x, uint16_t y, uint16_t w, uint16_t h, bool invert, bool mirror_y, bool pgm) {
   if (black)
     writeImage(black, x, y, w, h, invert, mirror_y, pgm);
 }
 
-void GxEPD2_371::writeImagePart(const uint8_t* black, const uint8_t* color, int16_t x_part, int16_t y_part, int16_t w_bitmap, int16_t h_bitmap,
-                                int16_t x, int16_t y, int16_t w, int16_t h, bool invert, bool mirror_y, bool pgm) {
+void GxEPD2_371::writeImagePart(const uint8_t* black, const uint8_t* color, uint16_t x_part, uint16_t y_part, uint16_t w_bitmap, uint16_t h_bitmap,
+                                uint16_t x, uint16_t y, uint16_t w, uint16_t h, bool invert, bool mirror_y, bool pgm) {
   if (black)
     writeImagePart(black, x_part, y_part, w_bitmap, h_bitmap, x, y, w, h, invert, mirror_y, pgm);
 }
 
-void GxEPD2_371::writeNative(const uint8_t* data1, const uint8_t* data2, int16_t x, int16_t y, int16_t w, int16_t h, bool invert, bool mirror_y, bool pgm) {
+void GxEPD2_371::writeNative(const uint8_t* data1, const uint8_t* data2, uint16_t x, uint16_t y, uint16_t w, uint16_t h, bool invert, bool mirror_y, bool pgm) {
   if (data1)
     writeImage(data1, x, y, w, h, invert, mirror_y, pgm);
 }
 
-void GxEPD2_371::drawImage(const uint8_t bitmap[], int16_t x, int16_t y, int16_t w, int16_t h, bool invert, bool mirror_y, bool pgm) {
+void GxEPD2_371::drawImage(const uint8_t bitmap[], uint16_t x, uint16_t y, uint16_t w, uint16_t h, bool invert, bool mirror_y, bool pgm) {
   writeImage(bitmap, x, y, w, h, invert, mirror_y, pgm);
   refresh(x, y, w, h);
   writeImage(bitmap, x, y, w, h, invert, mirror_y, pgm);
 }
 
-void GxEPD2_371::drawImagePart(const uint8_t bitmap[], int16_t x_part, int16_t y_part, int16_t w_bitmap, int16_t h_bitmap,
-                               int16_t x, int16_t y, int16_t w, int16_t h, bool invert, bool mirror_y, bool pgm) {
+void GxEPD2_371::drawImagePart(const uint8_t bitmap[], uint16_t x_part, uint16_t y_part, uint16_t w_bitmap, uint16_t h_bitmap,
+                               uint16_t x, uint16_t y, uint16_t w, uint16_t h, bool invert, bool mirror_y, bool pgm) {
   writeImagePart(bitmap, x_part, y_part, w_bitmap, h_bitmap, x, y, w, h, invert, mirror_y, pgm);
   refresh(x, y, w, h);
   writeImagePart(bitmap, x_part, y_part, w_bitmap, h_bitmap, x, y, w, h, invert, mirror_y, pgm);
 }
 
-void GxEPD2_371::drawImage(const uint8_t* black, const uint8_t* color, int16_t x, int16_t y, int16_t w, int16_t h, bool invert, bool mirror_y, bool pgm) {
+void GxEPD2_371::drawImage(const uint8_t* black, const uint8_t* color, uint16_t x, uint16_t y, uint16_t w, uint16_t h, bool invert, bool mirror_y, bool pgm) {
   writeImage(black, color, x, y, w, h, invert, mirror_y, pgm);
   refresh(x, y, w, h);
   writeImage(black, color, x, y, w, h, invert, mirror_y, pgm);
 }
 
-void GxEPD2_371::drawImagePart(const uint8_t* black, const uint8_t* color, int16_t x_part, int16_t y_part, int16_t w_bitmap, int16_t h_bitmap,
-                               int16_t x, int16_t y, int16_t w, int16_t h, bool invert, bool mirror_y, bool pgm) {
+void GxEPD2_371::drawImagePart(const uint8_t* black, const uint8_t* color, uint16_t x_part, uint16_t y_part, uint16_t w_bitmap, uint16_t h_bitmap,
+                               uint16_t x, uint16_t y, uint16_t w, uint16_t h, bool invert, bool mirror_y, bool pgm) {
   writeImagePart(black, color, x_part, y_part, w_bitmap, h_bitmap, x, y, w, h, invert, mirror_y, pgm);
   refresh(x, y, w, h);
   writeImagePart(black, color, x_part, y_part, w_bitmap, h_bitmap, x, y, w, h, invert, mirror_y, pgm);
 }
 
-void GxEPD2_371::drawNative(const uint8_t* data1, const uint8_t* data2, int16_t x, int16_t y, int16_t w, int16_t h, bool invert, bool mirror_y, bool pgm) {
+void GxEPD2_371::drawNative(const uint8_t* data1, const uint8_t* data2, uint16_t x, uint16_t y, uint16_t w, uint16_t h, bool invert, bool mirror_y, bool pgm) {
   writeNative(data1, data2, x, y, w, h, invert, mirror_y, pgm);
   refresh(x, y, w, h);
   writeNative(data1, data2, x, y, w, h, invert, mirror_y, pgm);
@@ -188,14 +185,14 @@ void GxEPD2_371::refresh(bool partial_update_mode) {
   }
 }
 
-void GxEPD2_371::refresh(int16_t x, int16_t y, int16_t w, int16_t h) {
+void GxEPD2_371::refresh(uint16_t x, uint16_t y, uint16_t w, uint16_t h) {
   if (_initial_refresh) return refresh(false); // initial update needs be full update
   x -= x % 8;              // byte boundary
   w -= x % 8;              // byte boundary
-  int16_t x1 = x < 0 ? 0 : x; // limit
-  int16_t y1 = y < 0 ? 0 : y; // limit
-  int16_t w1 = x + w < int16_t(WIDTH) ? w : int16_t(WIDTH) - x; // limit
-  int16_t h1 = y + h < int16_t(HEIGHT) ? h : int16_t(HEIGHT) - y; // limit
+  uint16_t x1 = x < 0 ? 0 : x; // limit
+  uint16_t y1 = y < 0 ? 0 : y; // limit
+  uint16_t w1 = x + w < uint16_t(WIDTH) ? w : uint16_t(WIDTH) - x; // limit
+  uint16_t h1 = y + h < uint16_t(HEIGHT) ? h : uint16_t(HEIGHT) - y; // limit
   w1 -= x1 - x;
   h1 -= y1 - y;
   if (!_using_partial_mode) _Init_Part();
@@ -211,8 +208,7 @@ void GxEPD2_371::powerOff() {
 
 void GxEPD2_371::hibernate() {
   _PowerOff();
-  if (_rst >= 0)
-  {
+  if (_rst >= 0) {
     _writeCommand(0x07); // deep sleep
     _writeData(0xA5);    // check code
     _hibernating = true;
@@ -234,8 +230,7 @@ void GxEPD2_371::_setPartialRamArea(uint16_t x, uint16_t y, uint16_t w, uint16_t
 }
 
 void GxEPD2_371::_PowerOn() {
-  if (!_power_is_on)
-  {
+  if (!_power_is_on) {
     _writeCommand(0x04);
     _waitWhileBusy("_PowerOn", power_on_time);
   }
@@ -243,8 +238,7 @@ void GxEPD2_371::_PowerOn() {
 }
 
 void GxEPD2_371::_PowerOff() {
-  if (_power_is_on)
-  {
+  if (_power_is_on) {
     _writeCommand(0x03);   // power off sequence
     _writeData(0x30);
     _writeCommand(0x02);   // power off
